Introduce MatchingVersion
The new method allow to retrieve what is the current matching version of
Gutenberg.

This method is useful to reuse the same logic to query for the version that
will actually be loaded.

It is an addition to the value returned by `loadMatching` to be able to know
the matching version in a different context.